@@ -1,5 +1,5 @@
 {
-  "name": "new-folder",
+  "name": "typescript-node",
   "version": "1.0.0",
   "description": "",
   "license": "MIT",
@@ -14,36 +14,43 @@
     "nodejs"
   ],
   "scripts": {
-    "test": "gulp test"
+    "tsc": "tsc",
+    "tsc-watch": "tsc -w",
+    "build": "gulp build",
+    "start": "node build/src/index.js",
+    "setup": "npm install & typings install",
+    "test": "gulp test",
+    "watch": "nodemon --watch build/src build/src/index.js"
   },
   "dependencies": {
-    "boom": "^3.0.0",
+    "bcryptjs": "^2.3.0",
+    "boom": "^3.2.2",
     "fs": "0.0.2",
-    "good": "^6.6.0",
-    "good-console": "^5.3.1",
-    "hapi": "^13.0.0",
-    "hapi-swagger": "^4.2.1",
-    "inert": "^3.2.0",
-    "joi": "^7.0.1",
-    "moment": "^2.12.0",
-    "mongodb": "^2.1.7",
-    "node-uuid": "^1.4.7",
+    "good": "^6.6.3",
+    "good-console": "^5.3.2",
+    "hapi": "^13.5.0",
+    "hapi-auth-jwt2": "^7.0.1",
+    "hapi-swagger": "^4.3.1",
+    "inert": "^3.2.1",
+    "joi": "^8.4.2",
+    "jsonwebtoken": "^7.1.7",
+    "mongoose": "^4.6.1",
+    "nconf": "^0.8.4",
     "path": "^0.12.7",
-    "vision": "^4.0.1"
+    "vision": "^4.1.0"
   },
   "devDependencies": {
     "chai": "^3.5.0",
-    "gulp": "^3.9.0",
-    "gulp-help": "^1.6.0",
-    "gulp-istanbul": "^0.10.3",
-    "gulp-mocha": "^2.1.3",
-    "gulp-nodemon": "^2.0.4",
+    "gulp": "^3.9.1",
+    "gulp-env": "^0.4.0",
+    "gulp-mocha": "^2.2.0",
     "gulp-rimraf": "^0.2.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-tslint": "^3.1",
-    "gulp-typescript": "^2.8.3",
-    "tslint": "^3.5.0",
-    "typescript": "^1.7.3"
+    "gulp-shell": "^0.5.2",
+    "gulp-tslint": "^4.3.4",
+    "mocha": "^3.0.0",
+    "nodemon": "^1.10.2",
+    "tslint": "^3.13.0",
+    "typescript": "^2.0.3"
   },
   "engines": {
     "node": ">=4.1.1"
