@@ -1,31 +1,32 @@
-import {IPlugin, IPluginInfo} from '../interfaces'
-import * as Hapi from 'hapi'
-const Inert = require('inert');
-const Vision = require('vision');
-const HapiSwagger = require('hapi-swagger');
+import {IPlugin, IPluginInfo} from "../interfaces";
+import * as Hapi from "hapi";
 
 export default (): IPlugin => {
     return {
         register: (server: Hapi.Server) => {
             server.register([
-                Inert,
-                Vision,
+                require('inert'),
+                require('vision'),
                 {
-                    register: HapiSwagger,
+                    register: require('hapi-swagger'),
                     options: {
                         info: {
                             title: 'Task Api',
-                            description: 'Simple Task Api.',
+                            description: 'Task Api Documentation',
                             version: '1.0'
                         },
                         tags: [
                             {
                                 'name': 'tasks',
                                 'description': 'Api tasks interface.'
+                            },
+                            {
+                                'name': 'users',
+                                'description': 'Api users interface.'
                             }
                         ],
                         enableDocumentation: true,
-                        documentationPath: '/documentation'
+                        documentationPath: '/docs'
                     }
                 }
             ]
@@ -41,5 +42,5 @@ export default (): IPlugin => {
                 version: "1.0.0"
             };
         }
-    }
+    };
 };
