@@ -1,83 +1,87 @@
-'use strict'
+'use strict';
 
-let gulp = require('gulp')
-let rimraf = require('gulp-rimraf')
-let tsc = require('gulp-typescript')
-let sourcemaps = require('gulp-sourcemaps')
-let tslint = require('gulp-tslint')
-let nodemon = require('gulp-nodemon')
-var mocha = require('gulp-mocha')
-var istanbul = require('gulp-istanbul')
-
-// /*  Variables */
-let tsProject = tsc.createProject('tsconfig.json')
-let sourceFiles = 'src/**/*.ts'
-let testFiles = 'test/**/*.ts'
-let outDir = require('./tsconfig.json').compilerOptions.outDir
-let entryPoint = './build/src/server.js'
+const gulp = require('gulp');
+const rimraf = require('gulp-rimraf');
+const tslint = require('gulp-tslint');
+const mocha = require('gulp-mocha');
+const shell = require('gulp-shell');
+const env = require('gulp-env');
 
 /**
  * Remove build directory.
  */
-gulp.task('clean', function() {
-    return gulp.src(outDir, { read: false })
-        .pipe(rimraf())
-})
-
+gulp.task('clean', function () {
+  return gulp.src(outDir, { read: false })
+    .pipe(rimraf());
+});
 
 /**
  * Lint all custom TypeScript files.
  */
 gulp.task('tslint', () => {
-    return gulp.src(sourceFiles)
-        .pipe(tslint())
-        .pipe(tslint.report('verbose'))
-})
+  return gulp.src('src/**/*.ts')
+    .pipe(tslint())
+    .pipe(tslint.report('prose'));
+});
 
 /**
- * Compile TypeScript sources and create sourcemaps in build directory.
+ * Compile TypeScript.
  */
-gulp.task('compile', ['clean'], () => {
-    let tsResult = gulp.src([sourceFiles, testFiles])
-        .pipe(sourcemaps.init())
-        .pipe(tsc(tsProject))
-    return tsResult.js
-        .pipe(sourcemaps.write('.'))
-        .pipe(gulp.dest(outDir))
-})
+
+function compileTS(args, cb) {
+  return exec(tscCmd + args, (err, stdout, stderr) => {
+    console.log(stdout);
+
+    if (stderr) {
+      console.log(stderr);
+    }
+    cb(err);
+  });
+}
+
+gulp.task('compile', shell.task([
+  'npm run tsc',
+]))
 
 /**
- * Watch for changes in TypeScript, HTML and CSS files.
+ * Watch for changes in TypeScript
  */
-gulp.task('watch', function() {
-    gulp.watch([sourceFiles], ['compile']).on('change', (e) => {
-        console.log('TypeScript file ' + e.path + ' has been changed. Compiling.')
-    })
-})
+gulp.task('watch', shell.task([
+  'npm run tsc-watch',
+]))
+/**
+ * Copy config files
+ */
+gulp.task('configs', (cb) => {
+  return gulp.src("src/configurations/*.json")
+    .pipe(gulp.dest('./build/src/configurations'));
+});
 
 /**
  * Build the project.
  */
-gulp.task('build', ['compile'], () => {
-    console.log('Building the project ...')
-})
+gulp.task('build', ['tslint', 'compile', 'configs'], () => {
+  console.log('Building the project ...');
+});
 
-gulp.task('test', ['build'], () => {
-    return gulp.src(['build/test/**/*.js'], { read: false })
-        .pipe(mocha({ reporter: 'list' }))
-        .once('error', () => {
-            process.exit(1);
-        })
-        .once('end', () => {
-            process.exit();
-        });
-})
+/**
+ * Run tests.
+ */
+gulp.task('test', ['build'], (cb) => {
+  const envs = env.set({
+    NODE_ENV: 'test'
+  });
 
-
-gulp.task('nodemon', ['build'], () => {
-    nodemon({
-        script: entryPoint,
-        env: { 'NODE_ENV': 'development' },
-        tasks: ['build']
+  gulp.src(['build/test/**/*.js'])
+    .pipe(envs)
+    .pipe(mocha())
+    .once('error', (error) => {
+      console.log(error);
+      process.exit(1);
     })
-})
+    .once('end', () => {
+      process.exit();
+    });
+});
+
+gulp.task('default', ['build']);
